+++ loefsys/core/__init__.py
@@ -1,0 +1,1 @@
+"""The core app contains all core mechanics required for loefsys."""
