+++ loefsys/core/apps.py
@@ -1,0 +1,7 @@
+"""Module containing the app definition for the core module."""
+from django.apps import AppConfig
+
+class CoreConfig(AppConfig):
+    """Configuration for the core."""
+
+    name = "loefsys.core"
