+++ loefsys/groups/urls.py
@@ -1,0 +1,11 @@
+"""Module containing the url definition of the sign up page."""
+
+from django.contrib.auth import views as auth_views
+from django.urls import path
+
+
+app_name = "groups"
+
+urlpatterns = [
+
+]
