--- loefsys/profile/__init__.py
@@ -1,4 +1,0 @@
-"""Django app for profiles.
-
-In this app, a profile is defined.
-"""
