--- loefsys/profile/apps.py
@@ -1,9 +1,0 @@
-"""Module defining the app configuration for the profile app."""
-
-from django.apps import AppConfig
-
-
-class ProfileConfig(AppConfig):
-    """Configuration class for the profile app."""
-
-    name = "loefsys.profile"
