--- loefsys/profile/static/input.css
@@ -1,10 +1,0 @@
-@import "tailwindcss";
-
-body {
-    font-family: Arial, sans-serif;
-    margin: 0;
-    padding: 0;
-    text-align: center;
-    background-color: white;
-    color: black;
-}
