--- loefsys/profile/templates/account_info/panel.html
@@ -1,8 +1,0 @@
-<div>
-  <h3>{{ key }}</h3>
-  {% if value == "" %}
-    <p class="italic">{{ key }} is not set.</p>
-  {% else %}
-    <p>{{ value }}</p>
-  {% endif %}
-</div>
