--- loefsys/profile/templates/password_reset/password_reset_confirm.html
@@ -1,64 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}
-Wachtwoord Aanpassen
-{% endblock %}
-
-{% block content %}
-<div class="flex flex-col h-screen">
-
-  <!-- Dark Blue Top Section -->
-  <section id="aanpassword-reset" class="bg-[#2972b3] p-8 flex flex-auto flex-col items-center">
-    <h1 class="text-6xl font-bold text-white pb-8 mt-4">Pas je wachtwoord aan.</h1>
-    <form method="post" class="bg-white p-8 rounded-3xl shadow-lg w-4/5">
-      {% csrf_token %}
-
-      <!-- Password Requirements Notice -->
-      <div class="mb-4 bg-yellow-100 border-l-4 border-yellow-500 text-yellow-700 p-4 rounded">
-        <p class="font-semibold">Wachtwoord vereisten:</p>
-        <ul class="list-disc pl-6">
-          <li>Minimaal 8 tekens</li>
-          <li>Minstens één hoofdletter</li>
-          <li>Minstens één kleine letter</li>
-          <li>Minstens één cijfer</li>
-          <li>Minstens één speciaal teken (bijv. !, @, #, etc.)</li>
-        </ul>
-      </div>
-
-      <!-- Password1 Field -->
-      <div class="mb-6">
-        <label for="{{ form.new_password1.id_for_label }}" class="block text-lg font-semibold mb-2">Nieuw wachtwoord</label>
-        <input type="password" name="{{ form.new_password1.name }}" id="{{ form.new_password1.id_for_label }}"
-               class="w-full text-5xl py-4 px-4 border rounded-lg" />
-        {% if form.new_password1.errors %}
-          <p class="text-red-500 text-sm mt-1">{{ form.new_password1.errors.0 }}</p>
-        {% endif %}
-      </div>
-
-      <!-- Password2 Field -->
-      <div class="mb-6">
-        <label for="{{ form.new_password2.id_for_label }}" class="block text-lg font-semibold mb-2">Bevestig wachtwoord</label>
-        <input type="password" name="{{ form.new_password2.name }}" id="{{ form.new_password2.id_for_label }}"
-               class="w-full text-5xl py-4 px-4 border rounded-lg" />
-        {% if form.new_password2.errors %}
-          <p class="text-red-500 text-sm mt-1">{{ form.new_password2.errors.0 }}</p>
-        {% endif %}
-      </div>
-
-      <!-- Submit Button -->
-      <button type="submit" class="p-10 border-2 rounded-3xl shadow-lg w-full bg-gray-200 hover:bg-gray-400 text-center mt-4 flex items-center justify-center">
-        <h1 class="font-bold text-5xl sm:text-5xl">Wachtwoord Bevestigen</h1>
-      </button>
-    </form>
-  </section>
-
-  <!-- Light Blue Bottom Section -->
-  <section id="back-to-login" class="bg-[#1BB1E6] p-8 flex flex-auto items-center flex-col">
-    <h1 class="text-4xl font-bold text-white pb-8 mt-4 text-center">TERUG NAAR INLOGGEN</h1>
-    <a href="{% url 'login' %}" class="btn btn-primary p-2 border-2 rounded-3xl shadow-lg w-4/5 bg-gray-200 hover:bg-gray-400 text-center">
-      <h1 class="font-bold text-3xl">Inloggen</h1>
-    </a>
-  </section>
-
-</div>
-{% endblock %}
