--- loefsys/profile/templates/password_reset/reset_password_email.html
@@ -1,12 +1,0 @@
-{% autoescape off %}
-Beste {{ user.display_name }},
-
-Je hebt aangegeven je wachtwoord te willen herstellen. Klik op de onderstaande link om een nieuw wachtwoord in te stellen:
-
-{{ protocol }}://{{ domain }}{% url 'password_reset_confirm' uidb64=uid token=token %}
-
-Als je dit niet hebt aangevraagd, kun je deze e-mail negeren.
-
-Met vriendelijke groet,
-Loefbijter
-{% endautoescape %}
