--- loefsys/profile/templates/password_reset/reset_password_subject.txt
@@ -1,1 +1,0 @@
-Aanpassen Loefbijter Wachtwoord
