--- loefsys/profile/templates/password_reset_complete.html
@@ -1,27 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}
-Wachtwoord Aangepast
-{% endblock %}
-
-{% block content %}
-<div class="flex flex-col h-screen">
-
-  <!-- Confirmation Message Section -->
-  <section class="bg-[#234193] flex flex-col p-8 flex-grow items-center">
-    <p class="text-5xl font-bold text-white mb-6">WACHTWOORD AANGEPAST</p>
-    <p class="text-white text-4xl mb-8 text-center max-w-md">
-      Je wachtwoord is succesvol aangepast. Je kan nu inloggen.
-    </p>
-  </section>
-
-  <!-- Light Blue Bottom Section -->
-  <section id="back-to-login" class="bg-[#1BB1E6] p-8 flex flex-auto items-center flex-col">
-    <p class="text-6xl font-bold text-white pb-8 mt-4 text-center">TERUG NAAR INLOGGEN</p>
-    <a href="{% url 'login' %}" class="btn btn-primary p-2 border-2 rounded-3xl shadow-lg w-4/5 bg-gray-200 hover:bg-gray-400 text-center">
-      <h1 class="font-bold text-3xl">Inloggen</h1>
-    </a>
-  </section>
-  {% include "menu.html" %}
-</div>
-{% endblock %}
