--- loefsys/profile/templates/password_reset_done.html
@@ -1,38 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}
-Wachtwoord Veranderen Aangevraagd
-{% endblock %}
-
-{% block content %}
-<div class="flex flex-col h-screen">
-
-  <!-- Dark Blue Top Section -->
-  <section id="password-reset" class="bg-[#234193] p-8 flex flex-auto flex-col items-center">
-    <p class="text-6xl font-bold text-white pb-8 mt-4">WACHTWOORD VERANDEREN AANGEVRAAGD</p> 
-
-    <p class="text-3xl font-bold text-white pb-8 mt-4">Kijk in je inbox</h2>
-    <br>
-    <p class="text-3xl text-white pb-8 mt-4l w-4/5">
-      We hebben je een email verstuurd met instructuies over hoe je je wachtwoord aanpast. 
-      Als het goed is zou je deze binnen enkele minuten in de inbox van het email adress van je loefbijter account moeten verschijnen.
-      <br>
-      <br>
-      Zie je deze mail niet verschijnen, chenk dan voor de zekerheid je spam mail. Anders kan je hieronder de mail opnieuw verzenden. 
-    </p>
-    <a href="{% url 'password_reset' %}" class="btn btn-primary p-2 border-2 rounded-3xl shadow-lg w-4/5 bg-gray-200 hover:bg-gray-400 text-center">
-      <h1 class="font-bold text-3xl">Opnieuw Proberen</h1>
-    </a>
-  </section>
-
-  <!-- Light Blue Bottom Section -->
-  <section id="back-to-login" class="bg-[#1BB1E6] p-8 flex flex-auto items-center flex-col">
-    <h1 class="text-4xl font-bold text-white pb-8 mt-4 text-center">TERUG NAAR INLOGGEN</h1>
-    <a href="{% url 'login' %}" class="btn btn-primary p-2 border-2 rounded-3xl shadow-lg w-4/5 bg-gray-200 hover:bg-gray-400 text-center">
-      <h1 class="font-bold text-3xl">Inloggen</h1>
-    </a>
-  </section>
-
-  {% include "menu.html" %}
-</div>
-{% endblock %}
