--- loefsys/profile/templates/request_password_reset.html
@@ -1,44 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}
-Wachtwoord Vergeten
-{% endblock %}
-
-{% block content %}
-<div class="flex flex-col h-screen">
-
-  <!-- Dark Blue Top Section -->
-  <section id="aanpassword-reset" class="bg-[#234193] p-8 flex flex-auto flex-col items-center">
-    <p class="text-6xl font-bold text-white pb-8 mt-4">WACHTWOORD WIJZIGING AANVRAGEN</p> 
-    <form method="post" class="bg-white p-8 rounded-3xl shadow-lg w-4/5">
-      {% csrf_token %}
-      
-      <!-- Email Field -->
-      <div class="mb-6">
-        <label for="{{ form.email.id_for_label }}" class="block text-lg font-semibold mb-2">Email</label>
-        <input type="email" name="{{ form.email.name }}" id="{{ form.email.id_for_label }}" 
-               value="{{ form.email.value|default_if_none:'' }}" 
-               class="w-full text-5xl py-4 px-5 border rounded-xl" />
-        {% if form.email.errors %}
-          <p class="text-red-500 text-sm mt-1">{{ form.email.errors.0 }}</p>
-        {% endif %}
-      </div>
-
-      <!-- Submit Button -->
-      <button type="submit" class="p-10 border-2 rounded-3xl shadow-lg w-full bg-gray-200 hover:bg-gray-400 text-center mt-4 flex items-center justify-center">
-        <h1 class="font-bold text-5xl sm:text-5xl">Verstuur link</h1>
-      </button>
-    </form>
-  </section>
-
-  <!-- Light Blue Bottom Section -->
-  <section id="back-to-login" class="bg-[#1BB1E6] p-8 flex flex-auto items-center flex-col">
-    <p class="text-5xl font-bold text-white pb-8 mt-4 text-center">TERUG NAAR INLOGGEN</p>
-    <a href="{% url 'login' %}" class="btn btn-primary p-2 border-2 rounded-3xl shadow-lg w-4/5 bg-gray-200 hover:bg-gray-400 text-center">
-      <h1 class="font-bold text-3xl">Inloggen</h1>
-    </a>
-  </section>
-
-  {% include "menu.html" %}
-</div>
-{% endblock %}
