--- loefsys/profile/templates/reset_password_email.html
@@ -1,17 +1,0 @@
-{% autoescape off %}
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="UTF-8">
-</head>
-<body>
-  <p>Hallo {{ user.display_name }},</p>
-  <p>Je ontvangt deze mail omdat je een verzoek hebt ingediend om je wachtwoord aan te passen.</p>
-  <p>
-    Klik op de link hieronder om je wachtwoord te resetten:<br />
-    <a href="{{ protocol }}://{{ domain }}{{ url }}">{{ protocol }}://{{ domain }}{{ url }}</a>
-  </p>
-  <p>Als je dit niet hebt aangevraagd, negeer deze mail dan.</p>
-</body>
-</html>
-{% endautoescape %}
